Add multiple requests coordination method
